fixed subtraction and shifting

+++ asm_ops.c
@@ -1,13 +1,15 @@
 #include "asm_ops.h"
+#include <stdint.h>
 
-int cmp_less(int a, int b) { return a < b; };
+uint64_t cmp_less(uint64_t a, uint64_t b) { return a < b; };
 
-int cmp_great(int a, int b) { return a > b; };
+uint64_t cmp_great(uint64_t a, uint64_t b) { return a > b; };
 
-int cmp_eq(int a, int b) { return a == b; };
+uint64_t cmp_eq(uint64_t a, uint64_t b) { return a == b; };
 
-int cmp_neq(int a, int b) { return a != b; };
+uint64_t cmp_neq(uint64_t a, uint64_t b) { return a != b; };
 
-int cmp_leq(int a, int b) { return a <= b; };
+uint64_t cmp_leq(uint64_t a, uint64_t b) { return a <= b; };
 
-int cmp_geq(int a, int b) { return a >= b; };
+uint64_t cmp_geq(uint64_t a, uint64_t b) { return a >= b; };
+
