+++ asm_ops.h
@@ -1,11 +1,13 @@
+#include <stdint.h>
+
 #ifndef ASM_OPS_H
 #define ASM_OPS_H
 
-int cmp_less(int a, int b);
-int cmp_great(int a, int b);
-int cmp_eq(int a, int b);
-int cmp_neq(int a, int b);
-int cmp_leq(int a, int b);
-int cmp_geq(int a, int b);
+uint64_t cmp_less(uint64_t a, uint64_t b);
+uint64_t cmp_great(uint64_t a, uint64_t b);
+uint64_t cmp_eq(uint64_t a, uint64_t b);
+uint64_t cmp_neq(uint64_t a, uint64_t b);
+uint64_t cmp_leq(uint64_t a, uint64_t b);
+uint64_t cmp_geq(uint64_t a, uint64_t b);
 
 #endif
